+++ R/igraph.r
@@ -27,7 +27,8 @@
 #'   igraph1 <- gexf.to.igraph(gexf1)
 #'   gexf2 <- igraph.to.gexf(igraph1)
 #'   
-#'   # Now, let's do it with a layout!
+#'   # Now, let's do it with a layout! (although we can just use
+#'   # the one that comes with lesmiserables :))
 #'   pos <- igraph::layout_nicely(igraph1)
 #'   plot(
 #'     igraph.to.gexf(igraph1, nodesVizAtt = list(position=cbind(pos, 0))),
@@ -42,48 +43,24 @@ igraph.to.gexf <- function(igraph.obj, ...) {
   
   # Retrive elements from igraph object
   gdata    <- igraph::get.data.frame(g, what="both")
-  tmpedges <- gdata$edges
-  tmpnodes <- gdata$vertices
-  
+  edges <- gdata$edges
+  nodes <- cbind(id = 1L:igraph::vcount(g), gdata$vertices)
+
   # If nodes have no name
-  if (!("name" %in% colnames(tmpnodes))) 
-    tmpnodes <- data.frame(tmpnodes, name=1:nrow(tmpnodes))
+  namecol <- which(colnames(nodes) == "name")
+  if (!length(namecol)) 
+    nodes <- data.frame(nodes, label=nodes$id)
+  else
+    nodes$label <- nodes$name
   
-  # Nodes and edges list
-  # these steps are dangerous as nodes that are not connected are dropped, creating a mismatched number of rows
-#   nodes <- edge.list(tmpedges[,c(1,2)])
-#   edges <- nodes$edges
-#   nodes <- nodes$nodes
-  
-  nodes <- tmpnodes
-  # change name to label
-  names(nodes) <- 'label'
-  # add a column to nodes to hold the IDs
-  nodes$id <- as.integer(as.factor(nodes$label))
-  # put them in the right order
-  nodes <- nodes[, c('id', 'label')]
-  
-  ## replace from and to in edges with their IDs
-  # first create a named vector of ids
-  nodesID <- nodes$id
-  names(nodesID) <- gdata$vertices$name
-  
-  edges <- tmpedges
-  # now in the edges create the id columns
-  edges$source <- nodesID[edges$from]
-  edges$target <- nodesID[edges$to]
-  # keep just source and target
+
+  edges$source <- nodes$id[match(edges$from, nodes$label)]
+  edges$target <- nodes$id[match(edges$to, nodes$label)]
   edges <- edges[, c('source', 'target')]
-  
-  # Building nodes
-  if (length(tmpnodes)) {
-    nodes <-merge(tmpnodes,nodes, by.x="name", by.y="label")
-#    nodes <- nodes[,!(colnames(nodes) %in% c("label"))]
-  }
   
   # Nodes Attributes
   x <- igraph::list.vertex.attributes(g)
-  x <- x[!(x %in% c("label","color","size"))]
+  x <- x[!(x %in% c("label","color","size", "name"))]
   if (length(x))
     dots$nodesAtt <- subset(nodes, select=x)
   
@@ -91,7 +68,7 @@ igraph.to.gexf <- function(igraph.obj, ...) {
   x <- igraph::list.edge.attributes(g)
   x <- x[!(x %in% c("weight","color","edgesLabel","width"))]
   if (length(x)) 
-    dots$edgesAtt <- subset(tmpedges, select=x)
+    dots$edgesAtt <- subset(edges, select=x)
   
   # Edges Weights
   if (!length(dots$edgesWeight) && length(igraph::E(g)$weight))
@@ -101,14 +78,14 @@ igraph.to.gexf <- function(igraph.obj, ...) {
   
   # Nodes Viz att
   if (!length(dots$nodesVizAtt$color)) {
-    if (length(tmpnodes$color)) {
-      dots$nodesVizAtt$color <- tmpnodes$color
+    if (length(nodes$color)) {
+      dots$nodesVizAtt$color <- nodes$color
     } else 
       dots$nodesVizAtt$color <- NULL
   }
   
-  if (!length(dots$nodesVizAtt$size) && length(tmpnodes$size))
-    dots$nodesVizAtt$size <- tmpnodes$size
+  if (!length(dots$nodesVizAtt$size) && length(nodes$size))
+    dots$nodesVizAtt$size <- nodes$size
   
   positions <- igraph::graph_attr(g, "layout")
   
@@ -119,8 +96,8 @@ igraph.to.gexf <- function(igraph.obj, ...) {
     
   # Edges Viz att
   if (!length(dots$edgesVizAtt$color)) {
-    if (length(tmpedges$color)) {
-      dots$edgesVizAtt$color <- tmpedges$color
+    if (length(edges$color)) {
+      dots$edgesVizAtt$color <- edges$color
     } else
       dots$edgexVizAtt$color <- NULL
   }
@@ -137,8 +114,8 @@ igraph.to.gexf <- function(igraph.obj, ...) {
   # Building graph
   do.call(gexf, c(
     list(
-      nodes = nodes[,c("id","name")],
-      edges = edges
+      nodes = nodes[, c("id","name")],
+      edges = edges[, c("source", "target")]
       ), dots)
   )
 
@@ -171,7 +148,10 @@ gexf.to.igraph <- function(gexf.obj) {
     igraph::V(g2)$size <- x$value
   
   if (length(x <- g$nodesVizAtt$position))
-    g2 <- igraph::set_graph_attr(g2, "layout", unname(x))
+    g2 <- igraph::set_graph_attr(
+      g2, "layout", 
+      unname(as.matrix(x))
+      )
   
   # Nodes atts
   if (length(x <- g$nodes[, !(colnames(g$nodes) %in% c("id", "label"))]))